+++ tests/src/codegen/wasm/algo_array.rs
@@ -40,7 +40,7 @@ mod algo_array_tests {
         cov_mark::check_count!(wasm_codegen_emit_array_index_read, 24);
         cov_mark::check_count!(wasm_codegen_emit_array_index_write, 3);
         cov_mark::check_count!(wasm_codegen_emit_stack_prologue, 12);
-        cov_mark::check_count!(wasm_codegen_emit_stack_epilogue, 24);
+        cov_mark::check_count!(wasm_codegen_emit_stack_epilogue, 12);
         let test_name = "algo_array";
         let test_file_path = get_test_file_path(module_path!(), test_name);
         let source_code = std::fs::read_to_string(&test_file_path)